Move MetricRegistryUtils to prevent circular dependency.
